build: Update Dockerfile
This commit updates the base image for the build stage to
golang 1.21.

The base image for the runtime stage is updated to
gcr.io/distroless/static-debian12:nonroot. The previous base image
gcr.io/distroless/base is deprecated and is no longer updated.
Using the nonroot variant adds an extra layer of security.

Because the container should be run as an unpriveleged user the
default listening port is changed to 1053.

To make container builds as fast as possible the test run is
removed from the Dockerfile. Tests should run as part of the CI.